refactor(docker-tests): audit and fix test module placement
- Fix docker_tests_inner.rs feature gate: docker-tests-eth → docker-tests-ordermatch
  (module contains cross-chain ordermatching tests, not ETH-only tests)

- Split tendermint_tests.rs: extract cross-chain swap tests to tendermint_swap_tests.rs
  - swap_nucleus_with_doc (NUCLEUS <-> DOC)
  - swap_nucleus_with_eth (NUCLEUS <-> ETH)
  - swap_doc_with_iris_ibc_nucleus (DOC <-> IRIS-IBC-NUCLEUS)

- tendermint_swap_tests.rs gated by: docker-tests-tendermint + docker-tests-eth
  (requires both Tendermint and ETH docker environments)

- Update mod.rs comments to accurately describe module contents

- Mark audit task complete in plan file
